Implement image_size_with_srk()
This allows us to use the nboot command in uboot while using secure boot.
The nboot command reads in only the kernel image found in nand as
opposed to reading the whole nand partition (saving about 20 seconds!).
When secure boot is activated, however, we must also read in the SRK
data, which is located just past the kernel image.